Extracted rootview-specific shadowview logic into new class
Summary:It was hard to understand which parts of the shadowview API are designed to be called only on the root view, and which were applicable to any view.

This diff extracts rootview-specific logic out into a new RCTRootShadowView class.

Reviewed By: majak

Differential Revision: D3063905

fb-gh-sync-id: ef890cddfd7625fbd4bf5454314b441acdb03ac8
shipit-source-id: ef890cddfd7625fbd4bf5454314b441acdb03ac8

+++ React/Base/RCTPerformanceLogger.m
@@ -15,7 +15,7 @@
 #import "RCTProfile.h"
 
 static int64_t RCTPLData[RCTPLSize][2] = {};
-static int64_t RCTPLCookies[RCTPLSize] = {};
+static NSUInteger RCTPLCookies[RCTPLSize] = {};
 
 void RCTPerformanceLoggerStart(RCTPLTag tag)
 {
